remove brackets in file name post

+++ includes/response.hpp
@@ -1,21 +1,9 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   response.hpp                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: nidor <nidor@student.42.fr>                +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/08/05 14:17:50 by nidor             #+#    #+#             */
-/*   Updated: 2023/08/06 07:00:28 by nidor            ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
 #ifndef RESPONSE_HPP
 # define RESPONSE_HPP
 
 void		interruptResponse(int sock_clt, int sock_srv, const char *statusCode, const char *statusMessage);
 void        ft_split(std::string const &str, const char delim, std::vector<std::string> &out) ;
-int			send_header(int sock_clt, int sock_srv, int size, const char *path);
+int         send_header(int sock_clt, int sock_srv, long long int size, const char *path);
 void		print_request_header(int sock_clt, int sock_srv);
 void		serve_error_file(int sock_clt, int sock_srv);
 int			proceedResponse(int sock_clt, int sock_srv);
